feat: Implement onboarding feature with university and study program selection
- Added OnboardingRemoteDataSource and its implementation for fetching universities and study programs.
- Created OnboardingRepository and its implementation to handle user data updates and fetching universities/study programs.
- Introduced use cases for getting universities and study programs, and submitting onboarding data.
- Updated user model to include university and study program information.
- Enhanced onboarding page UI to allow user input for full name, avatar URL, GPA, university, and study program.
- Implemented Bloc for onboarding to manage state and handle events for loading universities, study programs, and submitting data.
- Created database schema for users, universities, and study programs.
- Added entities for University and StudyProgram to represent data structures.

+++ lib/presentation/widgets/academic_info_grid.dart
@@ -4,7 +4,7 @@ import '../../../../core/theme/app_pallete.dart';
 class AcademicInfoGrid extends StatelessWidget {
   final String programStudi;
   final int semester;
-  final int ipk;
+  final double ipk;
 
   const AcademicInfoGrid({
     super.key,
